Feat: #108 StatisticsRepository 리팩토링
- StatisticsRepository 인터페이스 분리
- StatisticsRepositoryImpl 구현체 분리
- RepositoryModule에서 @Binds를 @Provides로 변경
